+++ qbittorrent/rootfs/etc/cont-init.d/00-folders.sh
@@ -19,8 +19,8 @@ if [ -f /homeassistant/addons_config/qBittorrent/qBittorrent.conf ] && [ ! -f /h
     cp -rnp /homeassistant/addons_config/qBittorrent/* /config/qBittorrent/ &> /dev/null || true
     if [ -d /config/qBittorrent/addons_config ]; then rm -r /config/qBittorrent/addons_config; fi
     if [ -d /config/qBittorrent/qBittorrent ]; then rm -r /config/qBittorrent/qBittorrent; fi
-    echo "Files moved to /addon_configs/$HOSTNAME/openvpn" > /homeassistant/addons_config/qBittorrent/migrated
-    bashio::log.yellow "... moved files from /config/addons_config/qBittorrent to /addon_configs/$HOSTNAME/qBitorrent (must be accessed with my Filebrowser addon)"
+    echo "Files moved to /app_configs/$HOSTNAME/openvpn" > /homeassistant/addons_config/qBittorrent/migrated
+    bashio::log.yellow "... moved files from /config/addons_config/qBittorrent to /app_configs/$HOSTNAME/qBitorrent (must be accessed with my Filebrowser addon)"
     MIGRATED=true
 fi
 
@@ -37,7 +37,7 @@ fi
 if [ -d /homeassistant/openvpn ]; then
     if [ ! -f /homeassistant/openvpn/migrated ] && [ "$(ls -A /homeassistant/openvpn)" ]; then
         cp -rnf /homeassistant/openvpn/* /config/openvpn &> /dev/null || true
-        echo "Files moved to /addon_configs/$HOSTNAME/openvpn" > /homeassistant/openvpn/migrated
+        echo "Files moved to /app_configs/$HOSTNAME/openvpn" > /homeassistant/openvpn/migrated
     fi
 fi
 
@@ -45,14 +45,14 @@ fi
 if [ -f /homeassistant/addons_config/qbittorrent/config.yaml ] && [ ! -f /homeassistant/addons_config/qbittorrent/migrated ]; then
     cp -rnf /homeassistant/addons_config/qbittorrent/* /config/ &> /dev/null || true
     rm -r /homeassistant/addons_config/qbittorrent
-    bashio::log.yellow "... moved config.yaml from /config/addons_config/qbittorrent to /addon_configs/$HOSTNAME"
+    bashio::log.yellow "... moved config.yaml from /config/addons_config/qbittorrent to /app_configs/$HOSTNAME"
 fi
 
 # Move autoscript
 if [ -f /homeassistant/addons_autoscrips/qbittorrent.sh ]; then
     cp -rnf /homeassistant/addons_autoscrips/qbittorrent.sh /config/ &> /dev/null || true
     mv /homeassistant/addons_autoscrips/qbittorrent.sh /homeassistant/addons_autoscrips/qbittorrent.sh.bak
-    bashio::log.yellow "... moved qbittorrent.sh from /config/addons_autoscripts to /addon_configs/$HOSTNAME"
+    bashio::log.yellow "... moved qbittorrent.sh from /config/addons_autoscripts to /app_configs/$HOSTNAME"
 fi
 
 # Reboot post migration